Move Placement Requests controller
It bugged me having a 'placementRequests' folder without the 'placementRequestController'
in it
